analyzer: Do not use the package manager name as the package provider
While there are cases where the name of the (local) package manager equals
the name of the (remote) package provider / registry (Maven, NPM), there
are also other cases where the provider is named differently than the
manager (Bundler vs. RubyGems, PIP vs. PyPI). As a result, using a package
manager's toString() method is not always correct to set the provider of
a package (reference). So instead hard-code the provider in all cases to
emphasize it is not necessarily the same string as the package manager
name.