ReportTableModelMapper: Fix-up building the error description
The code for building the error description was touched by bf39794 which
intended to be a non-functional change. That change broke the error message
as `this` was referencing the ResolvableIssue instance while it actually
should reference the OrtIssue instance.

As result the error description was lacking the message from the OrtIssue
which inturn was thus missing in the `StaticHtml` and `Excel` report. Fix
this by referencing the correct instance with this.

Signed-off-by: Frank Viernau <frank.viernau@here.com>