CopyrightStatementsProcessor: Refactor matching the prefix
It is important to always use the longest matching prefix as
the result. The previous implementation was assuming that the
list of regexes used is sorted by the length of potential
matches of the respective regex. This makes managing the
array complex without need.

Remove this hidden assumption by always getting all matches and
taking the longest matching prefix as the result. As ordering
can be arbitrary now, use a more readable one.

Signed-off-by: Frank Viernau <frank.viernau@here.com>